+++ pkg/services/accesscontrol/acimpl/accesscontrol.go
@@ -1,4 +1,4 @@
-package ossaccesscontrol
+package acimpl
 
 import (
 	"context"
