+++ pkg/services/publicdashboards/api/common_test.go
@@ -16,6 +16,7 @@ import (
 	"github.com/grafana/grafana/pkg/infra/log"
 	"github.com/grafana/grafana/pkg/models"
 	"github.com/grafana/grafana/pkg/plugins"
+	"github.com/grafana/grafana/pkg/services/accesscontrol/acimpl"
 	"github.com/grafana/grafana/pkg/services/accesscontrol/database"
 	"github.com/grafana/grafana/pkg/services/accesscontrol/ossaccesscontrol"
 	"github.com/grafana/grafana/pkg/services/contexthandler/ctxkey"
@@ -63,7 +64,7 @@ func setupTestServer(
 	var err error
 	acService, err := ossaccesscontrol.ProvideService(cfg, database.ProvideService(db), rr)
 	require.NoError(t, err)
-	ac := ossaccesscontrol.ProvideAccessControl(cfg, acService)
+	ac := acimpl.ProvideAccessControl(cfg, acService)
 
 	// build mux
 	m := web.New()
